more data cleaning

+++ models/staging/raw/stg_raw__product.sql
@@ -9,8 +9,8 @@ source as (
 renamed as (
 
     select
-        CAST(products_id AS INT) AS products_id,
-        CAST(purchse_price AS FLOAT) AS purchse_price
+        CAST(products_id AS INT64) AS products_id,
+        CAST(purchse_price AS FLOAT64) AS purchse_price
 
     from source
